refactor(core): move functions from business to use_cases
